Introduce auxiliary peer address store
Introduce additional peer address store for `DiscoveryBehaviour` which
is an LRU cache of peers and their addresses that rest of the networking
uses to get dialing addresses for peer.

The addresses were stored in the DHT but Kademlia removes addresses it
knows about after a disconnection which leaves the peer in state where
`ProtocolController` is aware of the peer and may open a connection to
them which results in a call to `DiscoveryBehaviour` to collect known
addresses but since they were removed from Kademlia, the call results
in a dial failure.

To mitigate this, until the underlying problem with Kademlia is resolved,
store the addresses in a separate store which is queried when
a connection establishment attempt is made. The address store contains
all addresses that the peer might be reachable from and is purged as
dialing attempts fail. The address store is also backed up by an LRU
cache to prevent it from growing unreasonably large for long-running
nodes.